requirements.txt
dotenv
streamlit
google.generativeai
pypdf
open-cv-python-headless
python-docx
pillow